+++ tests/messages_data/rfc2822/example06.py
@@ -14,9 +14,9 @@ DATA = dict(
     html='',
     headers={'from': ('Mary Smith <mary@example.net>',), 'to': ('John Doe <jdoe@machine.example>',), 'reply-to': ('"Mary Smith: Personal Account" <smith@home.example>',), 'subject': ('Re: Saying Hello',), 'date': ('Fri, 21 Nov 1997 10:01:10 -0600',), 'message-id': ('<3456@example.net>',), 'in-reply-to': ('<1234@local.machine.example>',), 'references': ('<1234@local.machine.example>',)},
     attachments=[],
-    from_values=EmailAddress('Mary Smith', 'mary@example.net', 'Mary Smith <mary@example.net>'),
-    to_values=(EmailAddress('John Doe', 'jdoe@machine.example', 'John Doe <jdoe@machine.example>'),),
+    from_values=EmailAddress(name='Mary Smith', email='mary@example.net'),
+    to_values=(EmailAddress(name='John Doe', email='jdoe@machine.example'),),
     cc_values=(),
     bcc_values=(),
-    reply_to_values=(EmailAddress('Mary Smith: Personal Account', 'smith@home.example', 'Mary Smith: Personal Account <smith@home.example>'),),
+    reply_to_values=(EmailAddress(name='Mary Smith: Personal Account', email='smith@home.example'),),
 )